python-server/requirements.txt
Flask==2.2.2
kmeans_pytorch==0.3
matplotlib==3.7.0
numpy==1.23.5
pandas==1.5.3
torch==2.2.2+cu121
umap_learn==0.5.6
